refactor: ステージテストのロジックを分離
